Stop exporting the modules builder, encoder, and parser
