+++ src/error.rs
@@ -1,6 +1,7 @@
 use std::error::Error as StdError;
-use std::io::Error as StdIoError;
 use std::fmt;
+use std::io::Error as StdIoError;
+use std::result;
 
 use parser;
 use encoder;
@@ -20,6 +21,8 @@ pub enum Error {
     #[doc(hidden)]
     __Nonexhaustive,
 }
+
+pub type Result<T> = result::Result<T, Error>;
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
